Move zksync and utils to scripts dir
