+++ script/deploy/zksync/002_deploy_diamond.ts
