+++ script/deploy/zksync/006_deploy_receiver.ts
