+++ script/deploy/zksync/007_deploy_fee_collector.ts
