+++ script/deploy/zksync/008_deploy_service_fee_collector.ts
