+++ script/deploy/zksync/014_deploy_access_manager_facet.ts
@@ -1,7 +1,7 @@
 import { HardhatRuntimeEnvironment } from 'hardhat/types'
 import { DeployFunction } from 'hardhat-deploy/types'
 import { ethers, network } from 'hardhat'
-import { addOrReplaceFacets } from '../utils/diamond'
+import { addOrReplaceFacets } from '../../utils/diamond'
 import {
   diamondContractName,
   updateDeploymentLogs,
@@ -18,25 +18,26 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   const { deploy } = deployments
   const { deployer } = await getNamedAccounts()
 
-  const deployedFacet = await deploy('StandardizedCallFacet', {
+  const deployedFacet = await deploy('AccessManagerFacet', {
     from: deployer,
     log: true,
     skipIfAlreadyDeployed: true,
   })
 
-  const accessManagerFacet = await ethers.getContract('StandardizedCallFacet')
+  const accessManagerFacet = await ethers.getContract('AccessManagerFacet')
   const diamond = await ethers.getContract(diamondContractName)
 
   await addOrReplaceFacets([accessManagerFacet], diamond.address)
 
-  const isVerified = await verifyContract(hre, 'StandardizedCallFacet', {
+  const isVerified = await verifyContract(hre, 'AccessManagerFacet', {
     address: accessManagerFacet.address,
   })
 
-  await updateDeploymentLogs('StandardizedCallFacet', deployedFacet, isVerified)
+  await updateDeploymentLogs('AccessManagerFacet', deployedFacet, isVerified)
 }
 
 export default func
 
-func.id = 'deploy_standardized_call_facet'
-func.tags = ['DeployStandardizedCallFacet']
+func.id = 'deploy_access_manager_facet'
+func.tags = ['DeployAccessManagerFacet']
+func.dependencies = ['InitialFacets', diamondContractName, 'InitFacets']
