+++ script/deploy/zksync/019_deploy_lifuel_facet.ts
@@ -1,7 +1,7 @@
 import { HardhatRuntimeEnvironment } from 'hardhat/types'
 import { DeployFunction } from 'hardhat-deploy/types'
 import { ethers, network } from 'hardhat'
-import { addOrReplaceFacets } from '../utils/diamond'
+import { addOrReplaceFacets } from '../../utils/diamond'
 import {
   diamondContractName,
   updateDeploymentLogs,
@@ -18,26 +18,25 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   const { deploy } = deployments
   const { deployer } = await getNamedAccounts()
 
-  const deployedFacet = await deploy('AccessManagerFacet', {
+  const deployedFacet = await deploy('LIFuelFacet', {
     from: deployer,
     log: true,
     skipIfAlreadyDeployed: true,
   })
 
-  const accessManagerFacet = await ethers.getContract('AccessManagerFacet')
+  const facet = await ethers.getContract('LIFuelFacet')
   const diamond = await ethers.getContract(diamondContractName)
 
-  await addOrReplaceFacets([accessManagerFacet], diamond.address)
+  await addOrReplaceFacets([facet], diamond.address)
 
-  const isVerified = await verifyContract(hre, 'AccessManagerFacet', {
-    address: accessManagerFacet.address,
+  const isVerified = await verifyContract(hre, 'LIFuelFacet', {
+    address: facet.address,
   })
 
-  await updateDeploymentLogs('AccessManagerFacet', deployedFacet, isVerified)
+  await updateDeploymentLogs('LIFuelFacet', deployedFacet, isVerified)
 }
 
 export default func
 
-func.id = 'deploy_access_manager_facet'
-func.tags = ['DeployAccessManagerFacet']
-func.dependencies = ['InitialFacets', diamondContractName, 'InitFacets']
+func.id = 'deploy_lifuel_facet'
+func.tags = ['DeployLIFuelFacet']
