+++ script/deploy/zksync/9999_verify_all_facets.ts
