+++ script/utils/diamond.ts
